Remove not used, added some obsoletes just in case Add SimdUtf Add GlyphGeneration and rendering by the unicode character Refactor things, including CMakeLists

+++ GlyphAtlasSrc/Glyph/FontKey.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "PrimitiveTypes.h"
+
 enum TypeMask {
     Sdf,
     Msdf,
@@ -8,9 +10,9 @@ enum TypeMask {
 };
 
 struct FontKey {
-    uint8_t fontId{};
-    uint8_t faceIndex{};
-    long size{}; // FT_F26Dot6. Lazy to include the freeType library here.
+    uint8 fontId{};
+    uint8 faceIndex{};
+    int32 size{}; // FT_F26Dot6 aka signed long. Lazy to include the freeType library here.
 //    float sdfPadding; // not sure how to use these yet, and whether they should participate in the map keys.
 //    TypeMask mask;
